[FIXED JENKINS-23097] Set description during build as a build step.
This change allows the description of a build to be set as a part of
building. Setting the description early is useful if you have time
consuming builds.

Common description setter functionality is extracted to a helper
class in order to avoid code copy when implementing the new builder.
The common helper class also properly closes a reader after parsing a
log file.

The test builder is moved to a separate class in order to avoid code
copy when implementing more test classes. The test builder is now
compatible with jenkins-core 1.414 or later by defining a Descriptor,
mandatory since JENKINS-9687.

The parent version is bumped to 1.480.3 in order to compile against
Java 7.