Add support for multiline patterns
Changed the parseLog() function to read the entire log file, so that the
regular expression can match on multiple lines.

Our Jenkins server uses this to set the build description to be the same
as the Git commit message (using `git log --format=%B -n 1 HEAD`).

The log is read one line at a time and appended into a single string.
This is probably horribly inefficient, but I'm not very fluent in Java,
so it's good enough for our purposes.

It also reads the entire file before trying to match the pattern, but
on the other hand, the match is attempted only once (as opposed to once
for every line).

NOTE: Some tests failed for me, even with unmodified sources, so I
compiled without tests (`mvn install -Dmaven.test.skip=true`).