[JENKINS-72555] Only include HTML br tag when using HTML formatter
https://issues.jenkins.io/browse/JENKINS-72555 reports that a literal
'br' tag is included in the output when a build description already
exists and the description setter plugin is configured to append to the
build description.

Only add the 'br' tag if the HTML markup formatter is being used.
Other marketup formatters exist, but the plugin does not attempt to
adapt to those other markup formatters.

Intentionally does not create a runtime requirement on the OWASP markup
formatter plugin because this plugin should not require a formatter
plugin.