Require Jenkins 2.479 and Jakarta EE 9
Jenkins 2.479.1 unclude Jakarta EE 9, Spring Security 6, and Eclipse
Jetty 12.  It supports plugins that use Java EE 8, Spring Security 5,
and Eclipse Jetty 10 with a compatibility layer.  We want to remove
that compatibility layer in the future (as much as we can).  Update the
plugin to use Jakarta EE 9, Spring Security 6, and Eclipse Jetty 12
without using the compatibility layer.

Testing done

Confirmed that the Seed Jenkins plugin does not use any of the methods
whose signatures are modified by this pull request.

Confirmed that automated tests pass